error handling for comments added

+++ TPO/src/main/java/com/tpo/TPO/controller/CommentController.java
@@ -1,6 +1,8 @@
 package com.tpo.TPO.controller;
 
 import com.tpo.TPO.entity.Comment;
+import com.tpo.TPO.exceptions.NoCommentFound;
+import com.tpo.TPO.exceptions.NoMatchUserException;
 import com.tpo.TPO.service.CommentService;
 import com.tpo.TPO.service.UserService;
 
@@ -8,6 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import java.util.Optional;
 
 import java.util.List;
 
@@ -38,18 +41,37 @@ public class CommentController {
     @PostMapping
     public ResponseEntity<Comment> postCommentsPostID(@PathVariable Integer postId, @RequestBody Comment comment,
             @RequestHeader("Authorization") String authorizationHeader) {
+
         Integer userId = userService.getIdfromToken(authorizationHeader);
+
         Comment createdComment = commentService.createComment(postId, comment, userId);
         return ResponseEntity.status(HttpStatus.CREATED).body(createdComment);
     }
 
     // Delete a specific comment of a post
     @DeleteMapping("/{commentId}")
-    public ResponseEntity<Void> deleteComment(@PathVariable Integer postId, @PathVariable Integer commentId) {
-        boolean deleted = commentService.deleteComment(postId, commentId);
-        if (!deleted) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+    public ResponseEntity<?> deleteComment(@PathVariable Integer postId, @PathVariable Integer commentId,
+            @RequestHeader("Authorization") String authorizationHeader) throws NoMatchUserException, NoCommentFound {
+
+        try {
+
+            Optional<Comment> commentToDelete = commentService.getComment(postId, commentId);
+            Comment commentToDeleteReal = commentToDelete.get();
+            Integer userIdcommentToDelete = commentToDeleteReal.getUserId();
+            Integer userId = userService.getIdfromToken(authorizationHeader);
+
+            if (userIdcommentToDelete == userId) {
+
+                boolean deleted = commentService.deleteComment(postId, commentId);
+
+                return ResponseEntity.ok().body(deleted);
+
+            }
+            throw new NoMatchUserException();
+
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e);
         }
-        return ResponseEntity.ok().build();
+
     }
 }
